+++ tests/aider_test.ts
@@ -1,5 +1,5 @@
 // This comment for vim-test plugin: Use denops' test() instead of built-in Deno.test()
-import { test } from "./testUtil.ts";
+import { test } from "./testRunner.ts";
 
 test("floating", "AiderRun should work", async (denops) => {
   await denops.cmd("AiderRun");
